Added functions to the calorietrackerViewController that allowed me to add new user calorie amounts,In addition worked on storyboard as I redid the layout of the app to better match the example in the sprint, the project is finished

+++ CalorieTracker/CalorieTracker/Model/User+Convenience.swift
@@ -16,12 +16,13 @@ extension User {
     }
     
     @discardableResult convenience init(time: Date = Date(),
-                                        calories: Int16,
+                                        calories: Int,
                                         context: NSManagedObjectContext = CoreDataStack.shared.mainContext
     ) {
         self.init(context: context)
         self.time = time
-        self.calories = calories
+        self.calories = Int64(calories)
+        
     }
 }
